requirements.txt
-i https://pypi.org/simple
aiosqlite==0.21.0; python_version >= '3.9'
annotated-doc==0.0.3; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.11.0; python_version >= '3.9'
certifi==2025.10.5; python_version >= '3.7'
cffi==2.0.0; python_full_version >= '3.9' and platform_python_implementation != 'PyPy'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.3.0; python_version >= '3.10'
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0; python_version >= '3.10'
ecdsa==0.19.1; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
email-validator==2.3.0
fastapi==0.121.0; python_version >= '3.8'
greenlet==3.2.4; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.16.0; python_version >= '3.8'
httptools==0.7.1
idna==3.11; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pycparser==2.23; implementation_name != 'PyPy'
pydantic[email]==2.12.3; python_version >= '3.9'
pydantic-core==2.41.4; python_version >= '3.9'
python-dotenv==1.2.1; python_version >= '3.9'
python-jose[cryptography]==3.5.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pyyaml==6.0.3
requests==2.32.5; python_version >= '3.9'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.44; python_version >= '3.7'
starlette==0.49.3; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn[standard]==0.38.0; python_version >= '3.9'
watchfiles==1.1.1
websockets==15.0.1
